Changed tifffile.imsave to tifffile.imwrite
On the 2025.2.18 version, tifffile removed the imsave method, which was deprecated, and is replaced
with imwrite. Tifffile instalations equal or greater to that version makes the software to fail.
By updating the methods we ensure software continuity.

Changelog: https://pypi.org/project/tifffile/ 2025.2.18
Remove deprecated imsave and TiffWriter.save function/method (breaking).